+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 4.3
 - Add MakeChooseCheckbox() function in Inputs.php
 - Add FoodServiceReminderOutput() function in Reminders.php
 - SQL COURSES table: Add DESCRIPTION column in Update.fnc.php
+- Add Description (TinyMCE input) to Course in Courses.php
 
 Changes in 4.2
 --------------